+++ zuora_client/models/put_subscription_resume_type.py
@@ -149,6 +149,10 @@ class PUTSubscriptionResumeType(object):
 
         self._collect = collect
 
+    @collect.deleter
+    def collect(self):
+        del self._collect
+
     @property
     def contract_effective_date(self):
         """Gets the contract_effective_date of this PUTSubscriptionResumeType.  # noqa: E501
@@ -240,6 +244,10 @@ class PUTSubscriptionResumeType(object):
         """
 
         self._invoice = invoice
+
+    @invoice.deleter
+    def invoice(self):
+        del self._invoice
 
     @property
     def invoice_collect(self):
@@ -403,6 +411,10 @@ class PUTSubscriptionResumeType(object):
         """
 
         self._run_billing = run_billing
+
+    @run_billing.deleter
+    def run_billing(self):
+        del self._run_billing
 
     @property
     def target_date(self):